Update the beam centre to be a double parameter, and perform an automatic estimation.

+++ uk.ac.diamond.scisoft.ncd.processing/src/uk/ac/diamond/scisoft/analysis/processing/operations/ncd/GratingFitModel.java
@@ -6,13 +6,13 @@ import org.eclipse.dawnsci.analysis.api.processing.model.OperationModelField;
 public class GratingFitModel extends AbstractOperationModel {
 
 	@OperationModelField(hint = "Enter the beam centre as a two comma separated values.", label = "Beam Centre")
-	private int[] beamCentre;
+	private double[] beamCentre;
 	
-	public int[] getBeamCentre() {
+	public double[] getBeamCentre() {
 		return this.beamCentre;
 	}
 	
-	public void setBeamCentre(int[] beamCentre) {
+	public void setBeamCentre(double[] beamCentre) {
 		firePropertyChange("beamCentre", this.beamCentre, this.beamCentre = beamCentre);
 	}
